+++ Sprint 7/mediam/mediam/urls.py
@@ -3,8 +3,7 @@ from django.urls import path, include
 
 urlpatterns = [
     path('', include('home.urls')),
-    path('', include('auth.urls')),
-    path('', include('perfil.urls')),
+    path('', include('user.urls')),
     path('', include('chat.urls')),
     path('', include('upload_image.urls')),
 ]
