--- Sprint 7/mediam/perfil/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class PerfilConfig(AppConfig):
-    name = 'perfil'
