--- Sprint 7/mediam/perfil/urls.py
@@ -1,8 +1,0 @@
-from django.contrib import admin
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    path('perfil/', views.perfil),
-    path('settings/', views.settings),
-]
