--- Sprint 7/mediam/perfil/views.py
@@ -1,8 +1,0 @@
-from django.shortcuts import render
-
-# Create your views here.
-def perfil(request):
-    return render(request, 'perfil/perfil.html')
-
-def settings(request):
-    return render(request, 'perfil/settings.html')
