+++ Sprint 7/mediam/user/urls.py
@@ -5,5 +5,7 @@ from . import views
 urlpatterns = [
     path('login/', views.login),
     path('register/', views.register),
-    path('reset_password/', views.reset_password)
+    path('reset_password/', views.reset_password),
+    path('perfil/', views.perfil),
+    path('settings/', views.settings),
 ]
